push extra backend

+++ backend/src/main/java/com/dut/backend/repository/ShoppingCartRepository.java
@@ -8,7 +8,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Integer> {
+public interface ShoppingCartRepository extends JpaRepository<ShoppingCart, Long> {
     List<ShoppingCart> findByAccountId(Long accountId);
     Optional<ShoppingCart> findByAccountIdAndItemId(Long accountId, Long itemId);
 }
